refactor(my-stories): extract story row and table hook
